requirements.txt
fastapi
pydantic
azure-monitor-opentelemetry